geometry swc: compute the heights of the truncated cones

+++ projects/geometry/include/neurite/geometry/types/truncated_cone.hpp
@@ -50,7 +50,7 @@ public:
         const double first_radius,
         const double second_radius,
         const neurite::geometry::vector3d& second,
-        const unsigned int height);
+        const double height);
 
 private:
     template<typename Archive>
@@ -98,8 +98,8 @@ public:
     void second(const neurite::geometry::vector3d& v);
     void second(const neurite::geometry::vector3d&& v);
 
-    unsigned int height() const;
-    void height(const unsigned int v);
+    double height() const;
+    void height(const double v);
 
 public:
     bool operator==(const truncated_cone& rhs) const;
@@ -119,7 +119,7 @@ private:
     double first_radius_;
     double second_radius_;
     neurite::geometry::vector3d second_;
-    unsigned int height_;
+    double height_;
 };
 
 } }
